test(viewmodels): relocate P0 ViewModel tests into per-feature dirs
Per Coverage_Roadmap §1.2, the P0 ViewModel test files now live at the
canonical per-feature paths so they sit alongside the modules they
exercise. The existing comprehensive suites (CatalogSearchViewModel,
HoldsViewModel, BookDetailViewModel) move from PalaceTests/ViewModels/
into PalaceTests/CatalogUI/, PalaceTests/Holds/, PalaceTests/Book/.

Removes an unwired duplicate Book/BookDetailViewModelTests.swift
(class BookDetailViewModelUnitTests, 44 tests) that overlapped the
canonical wired class BookDetailViewModelTests (85 tests).

Wire-up:
- new PalaceTests/Holds group added to PalaceTests parent group
- CatalogSearchViewModelTests.swift moved into CatalogUI group
- BookDetailViewModelTests.swift moved into Book group
- HoldsViewModelTests.swift placed in the new Holds group

All 175 tests under the three target classes pass via the swarm
verification command on iOS Simulator DF4A2A27.
