+++ PalaceTests/CatalogUI/CatalogSearchViewModelTests.swift
