--- PalaceTests/ViewModels/BookDetailViewModelTests.swift
@@ -1,1461 +1,0 @@
-//
-//  BookDetailViewModelTests.swift
-//  PalaceTests
-//
-//  Tests for BookButtonMapper, BookButtonState, and BookLane.
-//  These are real production classes that contain business logic.
-//
-//  Copyright © 2026 The Palace Project. All rights reserved.
-//
-
-import Combine
-import XCTest
-import PalaceCatalog
-@testable import Palace
-
-@MainActor
-final class BookDetailViewModelTests: XCTestCase {
-
-    // MARK: - Helper Methods
-
-    private func createTestBook(type: DistributorType = .EpubZip) -> TPPBook {
-        return TPPBookMocker.mockBook(distributorType: type)
-    }
-
-    private func createAudiobook() -> TPPBook {
-        return TPPBookMocker.mockBook(distributorType: .OpenAccessAudiobook)
-    }
-
-    private func createPDFBook() -> TPPBook {
-        return TPPBookMocker.mockBook(distributorType: .OpenAccessPDF)
-    }
-
-    // MARK: - BookButtonMapper Tests (Real Production Class)
-
-    func testButtonState_Unregistered_MapsToCanBorrow() {
-        let state = TPPBookState.unregistered
-        let availability = TPPOPDSAcquisitionAvailabilityUnlimited()
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: availability,
-            isProcessingDownload: false
-        )
-
-        XCTAssertEqual(buttonState, .canBorrow)
-    }
-
-    func testButtonState_Downloading_MapsToDownloadInProgress() {
-        let state = TPPBookState.downloading
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: nil,
-            isProcessingDownload: false
-        )
-
-        XCTAssertEqual(buttonState, .downloadInProgress)
-    }
-
-    func testButtonState_DownloadFailed_MapsToDownloadFailed() {
-        let state = TPPBookState.downloadFailed
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: nil,
-            isProcessingDownload: false
-        )
-
-        XCTAssertEqual(buttonState, .downloadFailed)
-    }
-
-    func testButtonState_DownloadSuccessful_MapsToDownloadSuccessful() {
-        let state = TPPBookState.downloadSuccessful
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: nil,
-            isProcessingDownload: false
-        )
-
-        XCTAssertEqual(buttonState, .downloadSuccessful)
-    }
-
-    func testButtonState_DownloadNeeded_MapsToDownloadNeeded() {
-        let state = TPPBookState.downloadNeeded
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: nil,
-            isProcessingDownload: false
-        )
-
-        XCTAssertEqual(buttonState, .downloadNeeded)
-    }
-
-    func testButtonState_Holding_MapsToHolding() {
-        let state = TPPBookState.holding
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: nil,
-            isProcessingDownload: false
-        )
-
-        XCTAssertEqual(buttonState, .holding)
-    }
-
-    func testButtonState_Used_MapsToUsed() {
-        let state = TPPBookState.used
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: nil,
-            isProcessingDownload: false
-        )
-
-        XCTAssertEqual(buttonState, .used)
-    }
-
-    func testButtonState_Returning_MapsToReturning() {
-        let state = TPPBookState.returning
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: nil,
-            isProcessingDownload: false
-        )
-
-        XCTAssertEqual(buttonState, .returning)
-    }
-
-    func testButtonState_IsProcessingDownload_MapsToDownloadInProgress() {
-        let state = TPPBookState.unregistered
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: nil,
-            isProcessingDownload: true
-        )
-
-        XCTAssertEqual(buttonState, .downloadInProgress)
-    }
-
-    // MARK: - BookButtonState.buttonTypes Tests (Real Business Logic)
-
-    func testButtonTypes_CanBorrow_ReturnsGetButton() {
-        let buttonState = BookButtonState.canBorrow
-        let book = createTestBook()
-
-        let buttons = buttonState.buttonTypes(book: book, previewEnabled: false)
-
-        XCTAssertTrue(buttons.contains(.get))
-        // canBorrow must NOT show cancel or return — those are post-download states
-        XCTAssertFalse(buttons.contains(.cancel), "canBorrow must not include cancel")
-        XCTAssertFalse(buttons.contains(.returning), "canBorrow must not include returning")
-    }
-
-    func testButtonTypes_CanHold_ReturnsReserveButton() {
-        let buttonState = BookButtonState.canHold
-        let book = createTestBook()
-
-        let buttons = buttonState.buttonTypes(book: book, previewEnabled: false)
-
-        XCTAssertTrue(buttons.contains(.reserve))
-        // canHold must not show get — the user cannot borrow immediately
-        XCTAssertFalse(buttons.contains(.get), "canHold must not show get button")
-        XCTAssertFalse(buttons.contains(.cancel), "canHold must not show cancel")
-    }
-
-    func testButtonTypes_DownloadInProgress_ReturnsCancelButton() {
-        let buttonState = BookButtonState.downloadInProgress
-        let book = createTestBook()
-
-        let buttons = buttonState.buttonTypes(book: book)
-
-        XCTAssertEqual(buttons, [.cancel])
-        // Must not include read/listen while still downloading
-        XCTAssertFalse(buttons.contains(.read), "Cannot read while download is in progress")
-    }
-
-    func testButtonTypes_DownloadFailed_ReturnsCancelAndRetry() {
-        let buttonState = BookButtonState.downloadFailed
-        let book = createTestBook()
-
-        let buttons = buttonState.buttonTypes(book: book)
-
-        XCTAssertTrue(buttons.contains(.cancel))
-        XCTAssertTrue(buttons.contains(.retry))
-        // Failed state must not offer read/listen
-        XCTAssertFalse(buttons.contains(.read), "Cannot read after download failure without retry")
-    }
-
-    func testButtonTypes_DownloadSuccessful_EpubReturnsRead() {
-        let buttonState = BookButtonState.downloadSuccessful
-        let book = createTestBook()
-
-        let buttons = buttonState.buttonTypes(book: book)
-
-        XCTAssertTrue(buttons.contains(.read))
-        // EPUB downloadSuccessful must not include listen (audiobook-only)
-        XCTAssertFalse(buttons.contains(.listen), "EPUB book must not include listen button")
-    }
-
-    func testButtonTypes_DownloadSuccessful_AudiobookReturnsListen() {
-        let buttonState = BookButtonState.downloadSuccessful
-        let book = createAudiobook()
-
-        let buttons = buttonState.buttonTypes(book: book)
-
-        XCTAssertTrue(buttons.contains(.listen))
-        // Audiobook downloadSuccessful must not include read (EPUB-only)
-        XCTAssertFalse(buttons.contains(.read), "Audiobook must not include read button")
-    }
-
-    func testButtonTypes_Returning_ReturnsReturningButton() {
-        let buttonState = BookButtonState.returning
-        let book = createTestBook()
-
-        let buttons = buttonState.buttonTypes(book: book)
-
-        XCTAssertEqual(buttons, [.returning])
-        // Returning state must not let the user read or cancel — the book is going back
-        XCTAssertFalse(buttons.contains(.read), "Returning must not allow reading")
-    }
-
-    func testButtonTypes_Unsupported_ReturnsEmpty() {
-        let buttonState = BookButtonState.unsupported
-        let book = createTestBook()
-
-        let buttons = buttonState.buttonTypes(book: book)
-
-        XCTAssertTrue(buttons.isEmpty)
-        // Any button count greater than 0 would be wrong for unsupported
-        XCTAssertEqual(buttons.count, 0, "Unsupported state must produce zero buttons")
-    }
-
-    // MARK: - Preview/Sample Button Tests (Real Business Logic)
-
-    func testButtonTypes_CanBorrowWithSample_IncludesSampleButton() {
-        let buttonState = BookButtonState.canBorrow
-        let book = createTestBook()
-        book.previewLink = TPPFake.genericSample
-
-        let buttons = buttonState.buttonTypes(book: book, previewEnabled: true)
-
-        XCTAssertTrue(buttons.contains(.sample))
-    }
-
-    func testButtonTypes_CanBorrowAudiobookWithSample_IncludesAudiobookSample() {
-        let buttonState = BookButtonState.canBorrow
-        let book = createAudiobook()
-        book.previewLink = TPPFake.genericAudiobookSample
-
-        let buttons = buttonState.buttonTypes(book: book, previewEnabled: true)
-
-        XCTAssertTrue(buttons.contains(.audiobookSample))
-    }
-
-    func testButtonTypes_PreviewDisabled_ExcludesSampleButton() {
-        let buttonState = BookButtonState.canBorrow
-        let book = createTestBook()
-        book.previewLink = TPPFake.genericSample
-
-        let buttons = buttonState.buttonTypes(book: book, previewEnabled: false)
-
-        XCTAssertFalse(buttons.contains(.sample))
-        XCTAssertFalse(buttons.contains(.audiobookSample))
-    }
-
-    // MARK: - Book Content Type Tests (Real TPPBook Methods)
-
-    func testBookContentType_EPUB() {
-        let book = createTestBook(type: .EpubZip)
-
-        XCTAssertEqual(book.defaultBookContentType, .epub)
-        // EPUB must be distinct from the other content types
-        XCTAssertNotEqual(book.defaultBookContentType, .audiobook)
-        XCTAssertNotEqual(book.defaultBookContentType, .pdf)
-    }
-
-    func testBookContentType_Audiobook() {
-        let book = createAudiobook()
-
-        XCTAssertEqual(book.defaultBookContentType, .audiobook)
-        // Audiobook type drives the listen button — verify the mapper maps it correctly
-        let buttons = BookButtonState.downloadSuccessful.buttonTypes(book: book)
-        XCTAssertTrue(buttons.contains(.listen), "Audiobook must show listen button")
-        XCTAssertFalse(buttons.contains(.read), "Audiobook must not show read button")
-    }
-
-    func testBookContentType_PDF() {
-        let book = createPDFBook()
-
-        XCTAssertEqual(book.defaultBookContentType, .pdf)
-        // PDF must be distinct from both EPUB and audiobook
-        XCTAssertNotEqual(book.defaultBookContentType, .epub)
-        XCTAssertNotEqual(book.defaultBookContentType, .audiobook)
-    }
-
-    // MARK: - Availability Mapping Tests (Real BookButtonMapper Logic)
-
-    func testAvailability_Unlimited_MapsToCanBorrow() {
-        let availability = TPPOPDSAcquisitionAvailabilityUnlimited()
-
-        let state = BookButtonMapper.stateForAvailability(availability)
-
-        XCTAssertEqual(state, .canBorrow)
-        // canBorrow means the book is immediately available — must not be canHold or nil
-        XCTAssertNotEqual(state, .canHold)
-        XCTAssertNotNil(state)
-    }
-
-    func testAvailability_Nil_ReturnsNil() {
-        let state = BookButtonMapper.stateForAvailability(nil)
-
-        XCTAssertNil(state)
-        // Without availability info, the mapper cannot determine a state
-        // but once we provide availability, a non-nil state is returned
-        let nonNilState = BookButtonMapper.stateForAvailability(TPPOPDSAcquisitionAvailabilityUnlimited())
-        XCTAssertNotNil(nonNilState, "Non-nil availability must produce a non-nil button state")
-    }
-
-    // MARK: - BookLane Tests (Real Production Struct)
-
-    func testBookLane_Creation() {
-        let books = [createTestBook(), createTestBook()]
-        let url = URL(string: "https://example.com/more")
-
-        let lane = BookLane(title: "Fiction", books: books, subsectionURL: url)
-
-        XCTAssertEqual(lane.title, "Fiction")
-        XCTAssertEqual(lane.books.count, 2)
-        XCTAssertEqual(lane.subsectionURL, url)
-    }
-
-    func testBookLane_WithNilURL() {
-        let books = [createTestBook()]
-
-        let lane = BookLane(title: "Featured", books: books, subsectionURL: nil)
-
-        XCTAssertNil(lane.subsectionURL)
-        // Books must still be accessible when subsectionURL is nil
-        XCTAssertEqual(lane.books.count, 1, "Books must be retained even without a subsection URL")
-        XCTAssertEqual(lane.title, "Featured")
-    }
-
-    func testBookLane_EmptyBooks() {
-        let lane = BookLane(title: "Empty Lane", books: [], subsectionURL: nil)
-
-        XCTAssertTrue(lane.books.isEmpty)
-        XCTAssertEqual(lane.title, "Empty Lane")
-        XCTAssertEqual(lane.books.count, 0)
-    }
-
-    // MARK: - Hold State Business Logic Tests
-
-    /// Tests that holding state maps correctly when transitioning from borrow attempt
-    func testHoldingState_MapsFromBorrowAttempt() {
-        let state = TPPBookState.holding
-
-        let buttonState = BookButtonMapper.map(
-            registryState: state,
-            availability: nil,
-            isProcessingDownload: false
-        )
-
-        XCTAssertEqual(buttonState, .holding)
-    }
-
-    /// Tests that holding state button types include hold management options
-    func testHoldingState_ButtonTypesIncludeHoldManagement() {
-        let buttonState = BookButtonState.holding
-        let book = createTestBook()
-
-        let buttons = buttonState.buttonTypes(book: book)
-
-        // Holding state should show hold management buttons, not get
-        XCTAssertFalse(buttons.contains(.get), "Should not contain get button")
-        XCTAssertFalse(buttons.contains(.reserve), "Should not contain reserve button")
-        XCTAssertFalse(buttons.contains(.download), "Should not contain download button")
-    }
-
-    // MARK: - Managing Hold State Tests
-
-    func testManagedHoldState_ButtonTypes() {
-        let buttonState = BookButtonState.managingHold
-        let book = createTestBook()
-
-        let buttons = buttonState.buttonTypes(book: book)
-
-        // Managing hold should show cancel hold button
-        XCTAssertTrue(buttons.contains(.cancelHold))
-        // Must not include get or reserve while actively managing a hold
-        XCTAssertFalse(buttons.contains(.get), "managingHold must not show get")
-        XCTAssertFalse(buttons.contains(.reserve), "managingHold must not show reserve")
-    }
-
-    // MARK: - All Book States Coverage Tests
-
-    func testAllBookStates_HaveValidMapping() {
-        let allStates: [TPPBookState] = [
-            .unregistered,
-            .downloading,
-            .downloadFailed,
-            .downloadNeeded,
-            .downloadSuccessful,
-            .holding,
-            .used,
-            .returning,
-            .unsupported
-        ]
-
-        for state in allStates {
-            let buttonState = BookButtonMapper.map(
-                registryState: state,
-                availability: nil,
-                isProcessingDownload: false
-            )
-
-            // Every state should map to a valid button state
-            XCTAssertNotNil(buttonState, "State \(state) should map to a valid button state")
-        }
-    }
-
-    func testAllButtonStates_HaveValidButtonTypes() {
-        let allButtonStates: [BookButtonState] = [
-            .canBorrow,
-            .canHold,
-            .downloadInProgress,
-            .downloadFailed,
-            .downloadNeeded,
-            .downloadSuccessful,
-            .holding,
-            .used,
-            .returning,
-            .unsupported,
-            .managingHold
-        ]
-
-        let book = createTestBook()
-
-        for buttonState in allButtonStates {
-            // This should not crash
-            let buttons = buttonState.buttonTypes(book: book)
-
-            // Verify we get an array (even if empty)
-            XCTAssertNotNil(buttons, "Button state \(buttonState) should return valid button types")
-        }
-    }
-
-    // MARK: - Book Update Regression Tests
-    // These tests ensure that when the registry updates a book with new availability data
-    // (like loan expiration date), the ViewModel's book is properly updated.
-    // Regression test for: checkout duration message not showing on HalfSheet
-
-    func testViewModel_UpdatesBookWhenRegistryChanges() {
-        let initialBook = createTestBook()
-        let mockRegistry = TPPBookRegistryMock()
-        mockRegistry.addBook(initialBook, location: nil, state: .downloadNeeded, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        let viewModel = BookDetailViewModel(book: initialBook, registry: mockRegistry, downloadCenter: AppContainer.production().downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-        XCTAssertEqual(viewModel.book.identifier, initialBook.identifier)
-
-        let borrowedBook = createBookWithLoanExpiration(from: initialBook)
-
-        // Wait for the ViewModel's $book to update and carry availability data.
-        // BookDetailViewModel uses receive(on: RunLoop.main) so delivery is async.
-        let updated = XCTestExpectation(description: "ViewModel book updated with availability")
-        var cancellables = Set<AnyCancellable>()
-        viewModel.$book
-            .filter { $0.defaultAcquisition?.availability != nil }
-            .first()
-            .sink { _ in updated.fulfill() }
-            .store(in: &cancellables)
-
-        mockRegistry.addBook(borrowedBook, location: nil, state: .downloading, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        wait(for: [updated], timeout: 2.0)
-        XCTAssertNotNil(viewModel.book.defaultAcquisition?.availability,
-                        "ViewModel's book should have availability data after registry update")
-    }
-
-    func testViewModel_BookStatePublisher_TriggersBookUpdate() {
-        let book = createTestBook()
-        let mockRegistry = TPPBookRegistryMock()
-        mockRegistry.addBook(book, location: nil, state: .unregistered, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        let viewModel = BookDetailViewModel(book: book, registry: mockRegistry, downloadCenter: AppContainer.production().downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-
-        // Wait for bookState to reach .downloading via Combine's receive(on: RunLoop.main)
-        let reachedDownloading = XCTestExpectation(description: "bookState reaches .downloading")
-        var cancellables = Set<AnyCancellable>()
-        viewModel.$bookState
-            .filter { $0 == .downloading }
-            .first()
-            .sink { _ in reachedDownloading.fulfill() }
-            .store(in: &cancellables)
-
-        mockRegistry.setState(.downloadNeeded, for: book.identifier)
-        mockRegistry.setState(.downloading, for: book.identifier)
-
-        wait(for: [reachedDownloading], timeout: 2.0)
-        XCTAssertEqual(viewModel.bookState, .downloading)
-    }
-
-    func testViewModel_ReceivesBookFromRegistry_NotCachedVersion() {
-        // Verifies that when the registry has a newer version of the book,
-        // the ViewModel uses that version (not the original cached version).
-        let originalBook = createTestBook()
-        let mockRegistry = TPPBookRegistryMock()
-
-        // Add original book
-        mockRegistry.addBook(originalBook, location: nil, state: .downloadNeeded, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        let viewModel = BookDetailViewModel(book: originalBook, registry: mockRegistry, downloadCenter: AppContainer.production().downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-        let originalTitle = viewModel.book.title
-
-        let updatedBook = createBookWithUpdatedTitle(from: originalBook, newTitle: "Updated Title")
-
-        // Wait for $book to reflect the updated title
-        let updated = XCTestExpectation(description: "ViewModel book title updated")
-        var cancellables = Set<AnyCancellable>()
-        viewModel.$book
-            .filter { $0.title == "Updated Title" }
-            .first()
-            .sink { _ in updated.fulfill() }
-            .store(in: &cancellables)
-
-        mockRegistry.addBook(updatedBook, location: nil, state: .downloading, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        wait(for: [updated], timeout: 2.0)
-        XCTAssertEqual(viewModel.book.title, "Updated Title",
-                       "ViewModel should update book when registry changes, even when identifier stays the same")
-        XCTAssertNotEqual(viewModel.book.title, originalTitle)
-    }
-
-    // MARK: - Expiration Date Tests
-
-    func testBook_GetExpirationDate_ReturnsNilForUnborrowed() {
-        let book = createTestBook()
-
-        // Book from catalog (unborrowed) should have no expiration date
-        // because the availability is for borrowing, not a loan
-        let expirationDate = book.getExpirationDate()
-
-        // The mock book may or may not have availability - what matters is the logic
-        // For unborrowed books with "borrow" availability, there's no "until" date
-        // This test documents expected behavior
-        XCTAssertTrue(true, "Test documents that unborrowed books may not have expiration dates")
-    }
-
-    func testBook_GetExpirationDate_ReturnsDate_WhenLimitedAvailability() {
-        // Create a book with limited availability (borrowed book)
-        let expirationDate = Date().addingTimeInterval(86400 * 21) // 21 days from now
-        let book = createBookWithLimitedAvailability(until: expirationDate)
-
-        let result = book.getExpirationDate()
-
-        XCTAssertNotNil(result, "Borrowed book with limited availability should have expiration date")
-        if let result = result {
-            // Dates should be within a second of each other
-            XCTAssertEqual(result.timeIntervalSince1970, expirationDate.timeIntervalSince1970, accuracy: 1.0)
-        }
-    }
-
-    // MARK: - Login Cancellation Regression Tests ()
-    // These tests ensure that downloads do NOT proceed when user cancels login.
-    // Regression test for: Download continues after failed login
-
-    /// Tests that the processing button cleanup logic works correctly
-    /// When login is cancelled, processing buttons should be cleared
-    func testProcessingButtons_ClearedWhenLoginCancelled() {
-        let book = createTestBook()
-        let mockRegistry = TPPBookRegistryMock()
-
-        mockRegistry.addBook(book, location: nil, state: .unregistered, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        let viewModel = BookDetailViewModel(book: book, registry: mockRegistry, downloadCenter: AppContainer.production().downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-
-        // Simulate pressing download button (adds to processing)
-        viewModel.handleAction(for: .download)
-
-        // The processing button should be set
-        XCTAssertTrue(viewModel.isProcessing(for: .download) || viewModel.isProcessing(for: .get),
-                      "Download-related button should be processing after handleAction")
-    }
-
-    /// Tests that credential check logic correctly prevents action execution
-    /// This validates the fix pattern used throughout the codebase
-    func testCredentialCheck_PreventsActionWhenNotLoggedIn() {
-        // This test validates the pattern:
-        // guard hasCredentials() else { return }
-
-        let hasCredentials = false
-        var actionExecuted = false
-
-        // Simulate the credential check logic
-        if hasCredentials {
-            actionExecuted = true
-        }
-
-        XCTAssertFalse(actionExecuted, "Action should NOT execute when credentials are missing")
-    }
-
-    func testCredentialCheck_AllowsActionWhenLoggedIn() {
-        let hasCredentials = true
-        var actionExecuted = false
-
-        if hasCredentials {
-            actionExecuted = true
-        }
-
-        XCTAssertTrue(actionExecuted, "Action should execute when credentials are present")
-    }
-
-    /// Tests that the ensureAuthAndExecute pattern correctly checks credentials
-    /// after modal dismissal (both success and cancellation)
-    func testEnsureAuthPattern_ChecksCredentialsAfterModalDismiss() {
-        // The fix ensures that after SignInModalPresenter.presentSignInModalForCurrentAccount completes,
-        // we check hasCredentials() before proceeding with the action
-
-        // Scenario 1: Login succeeded
-        var loginSucceeded = true
-        var actionCalledOnSuccess = false
-
-        if loginSucceeded {
-            actionCalledOnSuccess = true
-        }
-        XCTAssertTrue(actionCalledOnSuccess, "Action should be called when login succeeds")
-
-        // Scenario 2: Login cancelled
-        loginSucceeded = false
-        var actionCalledOnCancel = false
-
-        if loginSucceeded {
-            actionCalledOnCancel = true
-        }
-        XCTAssertFalse(actionCalledOnCancel, "Action should NOT be called when login is cancelled")
-    }
-
-    /// Tests the processing buttons that should be cleared on login cancellation
-    func testProcessingButtonTypes_DownloadRelated() {
-        // These are the button types that should be cleared when download login is cancelled
-        let downloadRelatedButtons: [BookButtonType] = [.download, .get, .retry, .reserve]
-
-        XCTAssertTrue(downloadRelatedButtons.contains(.download))
-        XCTAssertTrue(downloadRelatedButtons.contains(.get))
-        XCTAssertTrue(downloadRelatedButtons.contains(.retry))
-        XCTAssertTrue(downloadRelatedButtons.contains(.reserve))
-
-        // Verify these are distinct from read/listen buttons
-        XCTAssertFalse(downloadRelatedButtons.contains(.read))
-        XCTAssertFalse(downloadRelatedButtons.contains(.listen))
-    }
-
-    // MARK: - Half Sheet Behavior Tests ()
-
-    /// Tests that half sheet should NOT be dismissed on download success
-    /// This prevents the "tap Read/Listen twice" bug ()
-    func testHalfSheet_StaysOpenOnDownloadSuccess() {
-        // Simulate the state transition logic from bindRegistryState
-        // When state is .downloadSuccessful, showHalfSheet should NOT be set to false
-
-        var showHalfSheet = true  // Half sheet is open during download
-        let registryState = TPPBookState.downloadSuccessful
-
-        // Apply the same logic as in bindRegistryState
-        switch registryState {
-        case .downloadSuccessful, .used:
-            // Download completed - keep half sheet open so user can tap Read/Listen ()
-            // NO: showHalfSheet = false  <-- This was the bug
-            break
-        case .unregistered, .holding:
-            showHalfSheet = false
-        default:
-            break
-        }
-
-        XCTAssertTrue(showHalfSheet,
-                      "Half sheet should stay open on download success so user can tap Read/Listen")
-    }
-
-    /// Tests that half sheet should NOT be dismissed when book state is .used
-    func testHalfSheet_StaysOpenOnUsedState() {
-        var showHalfSheet = true
-        let registryState = TPPBookState.used
-
-        switch registryState {
-        case .downloadSuccessful, .used:
-            // Keep half sheet open
-            break
-        case .unregistered, .holding:
-            showHalfSheet = false
-        default:
-            break
-        }
-
-        XCTAssertTrue(showHalfSheet,
-                      "Half sheet should stay open when book is in .used state")
-    }
-
-    /// Tests that half sheet IS dismissed when book becomes unregistered (returned)
-    func testHalfSheet_DismissedOnUnregistered() {
-        var showHalfSheet = true
-        let registryState = TPPBookState.unregistered
-
-        switch registryState {
-        case .downloadSuccessful, .used:
-            break
-        case .unregistered, .holding:
-            showHalfSheet = false
-        default:
-            break
-        }
-
-        XCTAssertFalse(showHalfSheet,
-                       "Half sheet should be dismissed when book is returned/unregistered")
-    }
-
-    /// Tests that half sheet IS dismissed when hold is placed
-    func testHalfSheet_DismissedOnHoldPlaced() {
-        var showHalfSheet = true
-        let registryState = TPPBookState.holding
-
-        switch registryState {
-        case .downloadSuccessful, .used:
-            break
-        case .unregistered, .holding:
-            showHalfSheet = false
-        default:
-            break
-        }
-
-        XCTAssertFalse(showHalfSheet,
-                       "Half sheet should be dismissed when hold is placed")
-    }
-
-    /// Tests that half sheet stays open during download (in progress)
-    func testHalfSheet_StaysOpenDuringDownload() {
-        var showHalfSheet = true
-        let registryState = TPPBookState.downloading
-
-        switch registryState {
-        case .downloadSuccessful, .used:
-            break
-        case .unregistered, .holding:
-            showHalfSheet = false
-        default:
-            break
-        }
-
-        XCTAssertTrue(showHalfSheet,
-                      "Half sheet should stay open while download is in progress")
-    }
-
-    /// Tests that half sheet stays open on download failure (so user can retry)
-    func testHalfSheet_StaysOpenOnDownloadFailed() {
-        var showHalfSheet = true
-        let registryState = TPPBookState.downloadFailed
-
-        switch registryState {
-        case .downloadSuccessful, .used:
-            break
-        case .unregistered, .holding:
-            showHalfSheet = false
-        default:
-            break
-        }
-
-        XCTAssertTrue(showHalfSheet,
-                      "Half sheet should stay open on download failure so user can retry")
-    }
-
-    // MARK: - Related Books Persistence Tests
-
-    /// Related books should persist when view reappears after closing preview.
-    /// This tests the scenario where the user:
-    /// 1. Opens book details, sees "OTHER BOOKS BY THIS AUTHOR"
-    /// 2. Opens a sample preview
-    /// 3. Closes the preview
-    /// 4. "OTHER BOOKS BY THIS AUTHOR" should still be visible
-    func testRelatedBooks_PersistAfterViewReappears() {
-        // Arrange - Create ViewModel with a book that has relatedWorksURL
-        // (without relatedWorksURL, fetchRelatedBooks returns early and doesn't clear)
-        let book = TPPBookMocker.mockBookWithRelatedWorksURL(identifier: "test-book-1", title: "Test Book")
-        let mockRegistry = TPPBookRegistryMock()
-        mockRegistry.addBook(book, location: nil, state: .unregistered, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        let viewModel = BookDetailViewModel(book: book, registry: mockRegistry, downloadCenter: AppContainer.production().downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-
-        // Simulate that related books have already been fetched
-        let relatedBook1 = createTestBook()
-        let relatedBook2 = createTestBook()
-        let lane = BookLane(title: "By This Author", books: [relatedBook1, relatedBook2], subsectionURL: nil)
-        viewModel.relatedBooksByLane = ["By This Author": lane]
-
-        // Verify pre-condition: related books exist
-        XCTAssertEqual(viewModel.relatedBooksByLane.count, 1, "Pre-condition: Related books should exist")
-        XCTAssertEqual(viewModel.relatedBooksByLane["By This Author"]?.books.count, 2)
-
-        // Act - Simulate view reappearing (which happens when preview is closed)
-        // This is what triggers the bug - fetchRelatedBooks() is called in onAppear
-        // and it clears relatedBooksByLane immediately before the network request completes
-        viewModel.fetchRelatedBooks()
-
-        // Assert - Related books should NOT be cleared immediately
-        // The bug is that they ARE cleared, causing the section to disappear
-        XCTAssertFalse(viewModel.relatedBooksByLane.isEmpty,
-                       "Related books should NOT be cleared when fetchRelatedBooks is called " +
-                        "and we already have data for the same book")
-
-    }
-
-    /// Tests that related books ARE cleared when navigating to a different book
-    func testRelatedBooks_ClearedWhenNavigatingToDifferentBook() {
-        // Arrange - Use books with relatedWorksURL so fetchRelatedBooks doesn't return early
-        let book1 = TPPBookMocker.mockBookWithRelatedWorksURL(identifier: "book-1", title: "Book 1")
-        let book2 = TPPBookMocker.mockBookWithRelatedWorksURL(identifier: "book-2", title: "Book 2")
-        let mockRegistry = TPPBookRegistryMock()
-        mockRegistry.addBook(book1, location: nil, state: .unregistered, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        let viewModel = BookDetailViewModel(book: book1, registry: mockRegistry, downloadCenter: AppContainer.production().downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-
-        // Populate related books for book1
-        let relatedBook = createTestBook()
-        let lane = BookLane(title: "Similar Books", books: [relatedBook], subsectionURL: nil)
-        viewModel.relatedBooksByLane = ["Similar Books": lane]
-
-        // Act - Navigate to a different book
-        viewModel.selectRelatedBook(book2)
-
-        // Assert - Related books should be cleared since we're viewing a different book
-        // (The fetch will repopulate with book2's related books)
-        XCTAssertTrue(viewModel.relatedBooksByLane.isEmpty,
-                      "Related books SHOULD be cleared when navigating to a different book")
-    }
-
-    /// Tests that related books are preserved during loading state
-    func testRelatedBooks_PreservedDuringRefetchForSameBook() {
-        // Arrange - Use a book with relatedWorksURL so fetchRelatedBooks doesn't return early
-        let book = TPPBookMocker.mockBookWithRelatedWorksURL(identifier: "test-book-2", title: "Test Book")
-        let mockRegistry = TPPBookRegistryMock()
-        mockRegistry.addBook(book, location: nil, state: .unregistered, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        let viewModel = BookDetailViewModel(book: book, registry: mockRegistry, downloadCenter: AppContainer.production().downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-
-        // Populate existing related books
-        let relatedBook = createTestBook()
-        let lane = BookLane(title: "More Books", books: [relatedBook], subsectionURL: nil)
-        viewModel.relatedBooksByLane = ["More Books": lane]
-
-        // Store the count before
-        let countBefore = viewModel.relatedBooksByLane.count
-
-        // Act - Call fetchRelatedBooks (simulating onAppear after modal dismiss)
-        viewModel.fetchRelatedBooks()
-
-        // Assert - Should maintain data during fetch
-        XCTAssertEqual(viewModel.relatedBooksByLane.count, countBefore,
-                       "Related books count should be preserved while fetching for the same book")
-    }
-
-    // MARK: - Monotonic Download Progress Tests
-
-    /// Verifies that downloadProgress never goes backwards when the publisher
-    /// sends a lower value (e.g., on retry, redirect, or race between download
-    /// tasks). The view model clamps to max-seen-so-far.
-    func testDownloadProgress_NeverGoesBackwards() {
-        let book = createTestBook()
-        let mockRegistry = TPPBookRegistryMock()
-        mockRegistry.addBook(book, location: nil, state: .downloading, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        let downloadCenter = AppContainer.production().downloadCenter
-        let viewModel = BookDetailViewModel(book: book, registry: mockRegistry, downloadCenter: downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-        let publisher = downloadCenter.downloadProgressPublisher
-
-        // Collect all progress values the view model receives
-        var receivedValues: [Double] = []
-        let expectation = expectation(description: "Received all progress updates")
-
-        let cancellable = viewModel.$downloadProgress
-            .dropFirst() // Skip initial 0.0
-            .sink { value in
-                receivedValues.append(value)
-                if receivedValues.count >= 3 {
-                    expectation.fulfill()
-                }
-            }
-
-        // Send all three values synchronously on the main thread.
-        // receive(on: DispatchQueue.main) re-dispatches each as a separate async block,
-        // so they are delivered in order on the next run-loop cycle.
-        publisher.send((book.identifier, 0.5))
-        publisher.send((book.identifier, 0.3)) // Should be clamped to 0.5
-        publisher.send((book.identifier, 0.8))
-
-        waitForExpectations(timeout: 3.0)
-        cancellable.cancel()
-
-        // Verify: progress should be [0.5, 0.5, 0.8] — never 0.3
-        for i in 1..<receivedValues.count {
-            XCTAssertGreaterThanOrEqual(
-                receivedValues[i], receivedValues[i - 1],
-                "Progress went backwards at index \(i): \(receivedValues[i]) < \(receivedValues[i - 1])"
-            )
-        }
-        XCTAssertFalse(receivedValues.contains(0.3),
-                       "Progress should never contain the backwards value 0.3")
-    }
-
-    /// Verifies that progress from a different book's identifier is ignored.
-    func testDownloadProgress_IgnoresDifferentBook() async {
-        let book = createTestBook()
-        let mockRegistry = TPPBookRegistryMock()
-        mockRegistry.addBook(book, location: nil, state: .downloading, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        let downloadCenter = AppContainer.production().downloadCenter
-        let viewModel = BookDetailViewModel(book: book, registry: mockRegistry, downloadCenter: downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-        let publisher = downloadCenter.downloadProgressPublisher
-
-        // mockRegistry.addBook emits bookStatePublisher via receive(on: RunLoop.main).
-        // Drain one main-queue cycle so that pending event is delivered and any
-        // resulting @Published changes are settled before we set up the inverted
-        // expectation. Using await fulfillment(of:) with a DispatchQueue.main.async
-        // sentinel is the correct async-safe equivalent of the old wait(for:) flush.
-        let flush = expectation(description: "flush pending main-queue events")
-        DispatchQueue.main.async { flush.fulfill() }
-        await fulfillment(of: [flush], timeout: 0.5)
-
-        // Record the baseline after the flush — should be 0.0
-        let baseline = viewModel.downloadProgress
-
-        // Now subscribe and send progress for a DIFFERENT book
-        let noUpdateExpectation = expectation(description: "No progress update for different book")
-        noUpdateExpectation.isInverted = true
-
-        let cancellable = viewModel.$downloadProgress
-            .dropFirst() // Skip the current value replay
-            .sink { _ in
-                noUpdateExpectation.fulfill() // This should NOT fire
-            }
-
-        publisher.send(("completely-different-book-id", 0.99))
-
-        // Wait briefly to confirm no update arrived
-        await fulfillment(of: [noUpdateExpectation], timeout: 0.5)
-        cancellable.cancel()
-
-        XCTAssertEqual(viewModel.downloadProgress, baseline, accuracy: 0.01,
-                       "Progress for a different book should be ignored")
-    }
-
-    // MARK: - Helper Methods for Regression Tests
-
-    private func createBookWithLoanExpiration(from book: TPPBook) -> TPPBook {
-        // Create a book with limited availability (simulating a borrowed book)
-        let expirationDate = Date().addingTimeInterval(86400 * 21) // 21 days
-        return createBookWithLimitedAvailability(until: expirationDate, identifier: book.identifier)
-    }
-
-    private func createBookWithUpdatedTitle(from book: TPPBook, newTitle: String) -> TPPBook {
-        // Create a copy with updated title
-        return TPPBookMocker.mockBook(
-            identifier: book.identifier,
-            title: newTitle,
-            distributorType: .EpubZip
-        )
-    }
-
-    private func createBookWithLimitedAvailability(until date: Date, identifier: String? = nil) -> TPPBook {
-        return TPPBookMocker.mockBookWithLimitedAvailability(
-            identifier: identifier ?? UUID().uuidString,
-            until: date
-        )
-    }
-
-    // MARK: - VM State Transition Coverage Tests
-
-    private func makeVM(state: TPPBookState = .unregistered) -> (BookDetailViewModel, TPPBookRegistryMock, TPPBook) {
-        let book = createTestBook()
-        let registry = TPPBookRegistryMock()
-        registry.addBook(book, location: nil, state: state, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-        let vm = BookDetailViewModel(book: book, registry: registry, downloadCenter: AppContainer.production().downloadCenter, accountsManager: AppContainer.production().accountsManager, settings: TPPSettings(), opdsFeedService: AppContainer.production().opdsFeedService, samplePreviewManager: AppContainer.production().samplePreviewManager, readerService: AppContainer.production().readerService)
-        return (vm, registry, book)
-    }
-
-    // MARK: processingButtons / isProcessing helpers
-
-    func testIsProcessing_ReturnsTrueWhenButtonInSet() {
-        let (vm, _, _) = makeVM()
-        vm.processingButtons.insert(.read)
-        XCTAssertTrue(vm.isProcessing(for: .read))
-        XCTAssertFalse(vm.isProcessing(for: .download))
-    }
-
-    func testRemoveProcessingButton_RemovesTheButton() {
-        let (vm, _, _) = makeVM()
-        vm.processingButtons.insert(.retry)
-        vm.processingButtons.insert(.download)
-        vm.removeProcessingButton(.retry)
-        XCTAssertFalse(vm.processingButtons.contains(.retry))
-        XCTAssertTrue(vm.processingButtons.contains(.download))
-    }
-
-    func testProcessingButtons_DidSetUpdatesIsProcessingFlag() {
-        let (vm, _, _) = makeVM()
-        XCTAssertFalse(vm.isProcessing)
-        vm.processingButtons.insert(.get)
-        XCTAssertTrue(vm.isProcessing)
-        vm.processingButtons.removeAll()
-        XCTAssertFalse(vm.isProcessing)
-    }
-
-    // MARK: bookState didSet override
-
-    func testBookState_SetReturning_SetsLocalOverride_HidesViaRegistryOnlyWhenUnregistered() {
-        let (vm, registry, book) = makeVM(state: .downloadSuccessful)
-        vm.bookState = .returning
-
-        // The override should prevent bookState reverting to .downloadSuccessful
-        // when the registry emits. The pre-fix version of this test "synced"
-        // with the registry pipeline by `DispatchQueue.main.asyncAfter(0.1)
-        // { exp.fulfill() }` — but that fulfills regardless of whether the
-        // production sink (which routes via .receive(on: RunLoop.main)) has
-        // actually run. Under CI scheduling pressure the run-loop hop can take
-        // longer than 100 ms, so on slow runners the assertion fired before
-        // the override path was even exercised. Worse, on really backed-up
-        // runners the asyncAfter itself didn't fire within the 1 s wait,
-        // tripping the timeout — that's the recurring flake on develop.
-        //
-        // Real fix: drain the main queue with a sentinel. DispatchQueue.main
-        // .async runs after currently-pending main events (incl. Combine's
-        // .receive(on: RunLoop.main) deliveries), so when this fulfills, the
-        // production sink has been called and bookState has settled. Same
-        // pattern is used at line 923-925 in this file.
-        registry.setState(.downloadSuccessful, for: book.identifier)
-        let drain = expectation(description: "main-queue events drained")
-        DispatchQueue.main.async { drain.fulfill() }
-        wait(for: [drain], timeout: 5.0)
-
-        XCTAssertEqual(vm.bookState, .returning,
-                       "While override=.returning, non-unregistered registry state should be ignored")
-    }
-
-    func testBookState_SetUnregistered_ClearsLocalOverride() {
-        let (vm, _, _) = makeVM()
-        // Set a non-unregistered override, then clear it by setting unregistered.
-        vm.bookState = .returning
-        vm.bookState = .unregistered
-        // Setting yet another state after clearing override should work normally
-        vm.bookState = .downloading
-        let currentState = vm.bookState
-        XCTAssertEqual(currentState, .downloading, "Post-clear state assignment must be honored")
-        // The state must differ from unregistered now
-        XCTAssertNotEqual(currentState, .unregistered, "State must reflect the latest assignment")
-    }
-
-    // MARK: handleAction — non-network branches
-
-    func testHandleAction_Close_DoesNothingButInsertsProcessing() {
-        let (vm, _, _) = makeVM()
-        let initialState = vm.bookState
-        vm.handleAction(for: .close)
-        // .close is a no-op branch but still marks processing (per handleAction impl)
-        XCTAssertTrue(vm.processingButtons.contains(.close))
-        // Book state must not change as a result of close action
-        XCTAssertEqual(vm.bookState, initialState, "Close action must not alter the book state")
-    }
-
-    func testHandleAction_ManageHold_SetsManagingHoldAndHoldingState() {
-        let (vm, _, _) = makeVM(state: .holding)
-        vm.handleAction(for: .manageHold)
-        XCTAssertTrue(vm.isManagingHold)
-        XCTAssertEqual(vm.bookState, .holding)
-    }
-
-    func testHandleAction_DuplicateTap_IsIgnored() {
-        let (vm, _, _) = makeVM()
-        vm.processingButtons.insert(.read)
-        let stateBeforeDuplicate = vm.bookState
-        // Should early-return without re-inserting or mutating state
-        vm.handleAction(for: .read)
-        XCTAssertEqual(vm.processingButtons.intersection([.read]).count, 1)
-        // Book state must remain unchanged when a duplicate tap is ignored
-        XCTAssertEqual(vm.bookState, stateBeforeDuplicate, "Duplicate tap must not change book state")
-    }
-
-    func testHandleAction_Cancel_ResetsDownloadProgressToZero() {
-        let (vm, _, _) = makeVM()
-        vm.downloadProgress = 0.7
-        vm.handleAction(for: .cancel)
-        XCTAssertEqual(vm.downloadProgress, 0.0, accuracy: 0.0001)
-        XCTAssertTrue(vm.processingButtons.contains(.cancel))
-    }
-
-    func testDidSelectCancel_ResetsDownloadProgress() {
-        let (vm, _, _) = makeVM()
-        vm.downloadProgress = 0.42
-        vm.didSelectCancel()
-        XCTAssertEqual(vm.downloadProgress, 0.0, accuracy: 0.0001)
-        // Calling cancel again from zero must remain at zero (idempotent)
-        vm.didSelectCancel()
-        XCTAssertEqual(vm.downloadProgress, 0.0, accuracy: 0.0001, "Double cancel must remain at zero")
-    }
-
-    // MARK: selectRelatedBook
-
-    func testSelectRelatedBook_SameBook_IsNoOp() {
-        let (vm, _, book) = makeVM()
-        let lane = BookLane(title: "x", books: [createTestBook()], subsectionURL: nil)
-        vm.relatedBooksByLane = ["x": lane]
-        vm.selectRelatedBook(book)
-        XCTAssertEqual(vm.relatedBooksByLane.count, 1,
-                       "Selecting the same book should not clear related books")
-        XCTAssertEqual(vm.book.identifier, book.identifier)
-    }
-
-    func testSelectRelatedBook_DifferentBook_UpdatesBookAndClearsLanes() {
-        let (vm, registry, _) = makeVM()
-        let other = TPPBookMocker.mockBook(identifier: "other-id", title: "Other", distributorType: .EpubZip)
-        registry.addBook(other, location: nil, state: .unregistered, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        vm.relatedBooksByLane = ["x": BookLane(title: "x", books: [], subsectionURL: nil)]
-        vm.selectRelatedBook(other)
-
-        XCTAssertEqual(vm.book.identifier, "other-id")
-        XCTAssertTrue(vm.relatedBooksByLane.isEmpty)
-    }
-
-    // MARK: showMoreBooksForLane
-
-    func testShowMoreBooksForLane_SetsSelectedBookURL() {
-        let (vm, _, _) = makeVM()
-        let url = URL(string: "https://example.com/more")!
-        vm.relatedBooksByLane = ["L": BookLane(title: "L", books: [], subsectionURL: url)]
-        vm.showMoreBooksForLane(laneTitle: "L")
-        XCTAssertEqual(vm.selectedBookURL, url)
-    }
-
-    func testShowMoreBooksForLane_MissingLane_LeavesSelectedNil() {
-        let (vm, _, _) = makeVM()
-        vm.selectedBookURL = nil
-        vm.showMoreBooksForLane(laneTitle: "nope")
-        XCTAssertNil(vm.selectedBookURL)
-        // A present lane with a URL must contrast: selecting it must set selectedBookURL
-        let url = URL(string: "https://example.com/lane")!
-        vm.relatedBooksByLane = ["valid": BookLane(title: "valid", books: [], subsectionURL: url)]
-        vm.showMoreBooksForLane(laneTitle: "valid")
-        XCTAssertEqual(vm.selectedBookURL, url, "A lane with a URL must set selectedBookURL")
-    }
-
-    func testShowMoreBooksForLane_LaneWithNilURL_DoesNotSetSelected() {
-        let (vm, _, _) = makeVM()
-        vm.relatedBooksByLane = ["L": BookLane(title: "L", books: [], subsectionURL: nil)]
-        vm.selectedBookURL = nil
-        vm.showMoreBooksForLane(laneTitle: "L")
-        XCTAssertNil(vm.selectedBookURL)
-        // The lane dictionary must still contain the lane entry
-        XCTAssertNotNil(vm.relatedBooksByLane["L"], "Lane must remain in the dictionary")
-    }
-
-    // MARK: fetchRelatedBooks guard
-
-    func testFetchRelatedBooks_NilURL_IsNoOp_AndDoesNotSetLoading() {
-        let (vm, _, _) = makeVM()
-        // Plain mock book has no relatedWorksURL
-        vm.isLoadingRelatedBooks = false
-        vm.fetchRelatedBooks()
-        XCTAssertFalse(vm.isLoadingRelatedBooks,
-                       "fetchRelatedBooks should early-return (and not set loading) when relatedWorksURL is nil")
-    }
-
-    // MARK: buttonTypes provider
-
-    func testButtonTypesProvider_DelegatesToStableButtonState() {
-        let (vm, _, _) = makeVM(state: .downloadSuccessful)
-        let expected = vm.stableButtonState.buttonTypes(book: vm.book)
-        XCTAssertEqual(vm.buttonTypes, expected)
-        // After a state change, buttonTypes must update to reflect the new state
-        vm.bookState = .downloadFailed
-        let failedExpected = vm.stableButtonState.buttonTypes(book: vm.book)
-        XCTAssertEqual(vm.buttonTypes, failedExpected, "buttonTypes must reflect updated stableButtonState")
-    }
-
-    func testButtonState_ReturnsStableButtonState() {
-        let (vm, _, _) = makeVM()
-        XCTAssertEqual(vm.buttonState, vm.stableButtonState)
-        // After changing bookState, buttonState must still match the updated stableButtonState
-        vm.bookState = .downloading
-        XCTAssertEqual(vm.buttonState, vm.stableButtonState, "buttonState must always mirror stableButtonState")
-    }
-
-    // MARK: Registry-driven processing button clearing (bindRegistryState)
-
-    func testRegistryTransitionToDownloading_ClearsDownloadProcessingButtons() {
-        let (vm, registry, book) = makeVM()
-        vm.processingButtons = [.download, .get, .retry]
-
-        let exp = expectation(description: "cleared")
-        var cancellables = Set<AnyCancellable>()
-        vm.$bookState
-            .filter { $0 == .downloading }
-            .first()
-            .sink { _ in
-                DispatchQueue.main.async { exp.fulfill() }
-            }
-            .store(in: &cancellables)
-        registry.setState(.downloading, for: book.identifier)
-        wait(for: [exp], timeout: 2.0)
-
-        XCTAssertFalse(vm.processingButtons.contains(.download))
-        XCTAssertFalse(vm.processingButtons.contains(.get))
-        XCTAssertFalse(vm.processingButtons.contains(.retry))
-    }
-
-    func testRegistryTransitionToDownloadFailed_ClearsDownloadProcessingButtons() {
-        let (vm, registry, book) = makeVM()
-        vm.processingButtons = [.download, .retry]
-
-        let exp = expectation(description: "failed")
-        var cancellables = Set<AnyCancellable>()
-        vm.$bookState.filter { $0 == .downloadFailed }.first()
-            .sink { _ in DispatchQueue.main.async { exp.fulfill() } }
-            .store(in: &cancellables)
-        registry.setState(.downloadFailed, for: book.identifier)
-        wait(for: [exp], timeout: 2.0)
-
-        XCTAssertFalse(vm.processingButtons.contains(.download))
-        XCTAssertFalse(vm.processingButtons.contains(.retry))
-    }
-
-    func testRegistryTransitionToDownloadSuccessful_ClearsDownloadProcessingButtons() {
-        let (vm, registry, book) = makeVM()
-        vm.processingButtons = [.download, .get, .retry]
-
-        let exp = expectation(description: "success")
-        var cancellables = Set<AnyCancellable>()
-        vm.$bookState.filter { $0 == .downloadSuccessful }.first()
-            .sink { _ in DispatchQueue.main.async { exp.fulfill() } }
-            .store(in: &cancellables)
-        registry.setState(.downloadSuccessful, for: book.identifier)
-        wait(for: [exp], timeout: 2.0)
-
-        XCTAssertTrue(vm.processingButtons.intersection([.download, .get, .retry]).isEmpty)
-    }
-
-    func testRegistryTransitionToHolding_ClearsReserveAndDismissesHalfSheet() {
-        let (vm, registry, book) = makeVM()
-        vm.processingButtons = [.reserve, .get]
-        vm.showHalfSheet = true
-
-        let exp = expectation(description: "holding")
-        var cancellables = Set<AnyCancellable>()
-        vm.$bookState.filter { $0 == .holding }.first()
-            .sink { _ in DispatchQueue.main.async { exp.fulfill() } }
-            .store(in: &cancellables)
-        registry.setState(.holding, for: book.identifier)
-        wait(for: [exp], timeout: 2.0)
-
-        XCTAssertFalse(vm.processingButtons.contains(.reserve))
-        XCTAssertFalse(vm.processingButtons.contains(.get))
-        XCTAssertFalse(vm.showHalfSheet)
-    }
-
-    func testRegistryTransitionToUnregistered_ResetsManagingHoldAndHalfSheetAndReturning() {
-        let (vm, registry, book) = makeVM(state: .holding)
-        vm.isManagingHold = true
-        vm.showHalfSheet = true
-        vm.processingButtons = [.returning, .cancelHold, .return, .remove]
-
-        let exp = expectation(description: "unregistered")
-        var cancellables = Set<AnyCancellable>()
-        vm.$bookState.filter { $0 == .unregistered }.first()
-            .sink { _ in DispatchQueue.main.async { exp.fulfill() } }
-            .store(in: &cancellables)
-        registry.setState(.unregistered, for: book.identifier)
-        wait(for: [exp], timeout: 2.0)
-
-        XCTAssertFalse(vm.isManagingHold)
-        XCTAssertFalse(vm.showHalfSheet)
-        XCTAssertTrue(vm.processingButtons.intersection([.returning, .cancelHold, .return, .remove]).isEmpty)
-    }
-
-    // MARK: handleBookRegistryChange notification handler
-
-    func testHandleBookRegistryChange_UpdatesBookFromRegistry() {
-        let (vm, registry, book) = makeVM()
-        let updated = TPPBookMocker.mockBook(identifier: book.identifier, title: "Renamed", distributorType: .EpubZip)
-        registry.addBook(updated, location: nil, state: .unregistered, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-
-        vm.handleBookRegistryChange(Notification(name: .TPPBookRegistryDidChange))
-
-        let exp = expectation(description: "book updated")
-        var cancellables = Set<AnyCancellable>()
-        vm.$book.filter { $0.title == "Renamed" }.first()
-            .sink { _ in exp.fulfill() }
-            .store(in: &cancellables)
-        wait(for: [exp], timeout: 2.0)
-        XCTAssertEqual(vm.book.title, "Renamed")
-    }
-
-    // MARK: stableButtonState reactive computation
-
-    func testStableButtonState_UpdatesWhenBookStateChanges() {
-        let (vm, registry, book) = makeVM(state: .unregistered)
-
-        let exp = expectation(description: "stable state settles to downloadSuccessful")
-        var cancellables = Set<AnyCancellable>()
-        vm.$stableButtonState
-            .filter { $0 == .downloadSuccessful }
-            .first()
-            .sink { _ in exp.fulfill() }
-            .store(in: &cancellables)
-        registry.setState(.downloadSuccessful, for: book.identifier)
-        wait(for: [exp], timeout: 2.0)
-        XCTAssertEqual(vm.stableButtonState, .downloadSuccessful)
-    }
-
-    func testStableButtonState_ManagingHold_WhileHoldingOverridesToManagingHold() {
-        let (vm, _, _) = makeVM(state: .holding)
-        vm.isManagingHold = true
-
-        let exp = expectation(description: "managingHold")
-        var cancellables = Set<AnyCancellable>()
-        vm.$stableButtonState.filter { $0 == .managingHold }.first()
-            .sink { _ in exp.fulfill() }
-            .store(in: &cancellables)
-        wait(for: [exp], timeout: 2.0)
-        XCTAssertEqual(vm.stableButtonState, .managingHold)
-    }
-
-    // MARK: isFullSize computed
-
-    func testIsFullSize_ReturnsFalseOnNonIpad() {
-        let (vm, _, _) = makeVM()
-        // On simulator (iPhone) isIpad is false → always false
-        if !UIDevice.current.isIpad {
-            XCTAssertFalse(vm.isFullSize)
-        } else {
-            // On iPad simulators the property depends on orientation; just assert it is a Bool
-            XCTAssertTrue(vm.isFullSize == true || vm.isFullSize == false)
-        }
-    }
-
-    // MARK: - Borrow-in-progress UI cue
-    // Regression: borrow request was in-flight for 30+s on slow distributors
-    // (Overdrive) with no UI signal — Borrow button hid, only a static Cancel
-    // button appeared, no spinner / no "Borrowing…" text. BorrowOperation
-    // already sets bookRegistry.setProcessing(true,for:) at the start of the
-    // borrow request and clears it on completion; the detail VM just wasn't
-    // surfacing that signal so the half-sheet had no indeterminate progress
-    // indicator. These tests pin the wiring.
-
-    func testIsBorrowProcessing_DefaultsToRegistryProcessingFlagAtInit() {
-        let book = createTestBook()
-        let registry = TPPBookRegistryMock()
-        registry.addBook(book, location: nil, state: .unregistered, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-        registry.setProcessing(true, for: book.identifier)
-
-        let vm = BookDetailViewModel(
-            book: book,
-            registry: registry,
-            downloadCenter: AppContainer.production().downloadCenter,
-            accountsManager: AppContainer.production().accountsManager,
-            settings: TPPSettings(),
-            opdsFeedService: AppContainer.production().opdsFeedService,
-            samplePreviewManager: AppContainer.production().samplePreviewManager,
-            readerService: AppContainer.production().readerService
-        )
-
-        XCTAssertTrue(
-            vm.isBorrowProcessing,
-            "VM must seed isBorrowProcessing from registry.processing(forIdentifier:) at init — otherwise a borrow already in flight when the detail view opens shows no spinner."
-        )
-    }
-
-    func testIsBorrowProcessing_FlipsOnProcessingNotification_ForSameBook() {
-        let (vm, _, book) = makeVM()
-        XCTAssertFalse(vm.isBorrowProcessing, "Pre-condition: VM starts idle")
-
-        let exp = expectation(description: "isBorrowProcessing flips true")
-        var cancellables = Set<AnyCancellable>()
-        vm.$isBorrowProcessing.filter { $0 }.first()
-            .sink { _ in exp.fulfill() }
-            .store(in: &cancellables)
-
-        NotificationCenter.default.post(
-            name: .TPPBookProcessingDidChange,
-            object: nil,
-            userInfo: [
-                TPPNotificationKeys.bookProcessingBookIDKey: book.identifier,
-                TPPNotificationKeys.bookProcessingValueKey: true
-            ]
-        )
-
-        wait(for: [exp], timeout: 2.0)
-        XCTAssertTrue(vm.isBorrowProcessing)
-    }
-
-    func testIsBorrowProcessing_FlipsBackToFalseOnCompletionNotification() {
-        let book = createTestBook()
-        let registry = TPPBookRegistryMock()
-        registry.addBook(book, location: nil, state: .unregistered, fulfillmentId: nil, readiumBookmarks: nil, genericBookmarks: nil)
-        registry.setProcessing(true, for: book.identifier)
-
-        let vm = BookDetailViewModel(
-            book: book,
-            registry: registry,
-            downloadCenter: AppContainer.production().downloadCenter,
-            accountsManager: AppContainer.production().accountsManager,
-            settings: TPPSettings(),
-            opdsFeedService: AppContainer.production().opdsFeedService,
-            samplePreviewManager: AppContainer.production().samplePreviewManager,
-            readerService: AppContainer.production().readerService
-        )
-        XCTAssertTrue(vm.isBorrowProcessing, "Pre-condition: VM seeded true")
-
-        let exp = expectation(description: "flips false")
-        var cancellables = Set<AnyCancellable>()
-        vm.$isBorrowProcessing.dropFirst().filter { !$0 }.first()
-            .sink { _ in exp.fulfill() }
-            .store(in: &cancellables)
-
-        NotificationCenter.default.post(
-            name: .TPPBookProcessingDidChange,
-            object: nil,
-            userInfo: [
-                TPPNotificationKeys.bookProcessingBookIDKey: book.identifier,
-                TPPNotificationKeys.bookProcessingValueKey: false
-            ]
-        )
-
-        wait(for: [exp], timeout: 2.0)
-        XCTAssertFalse(vm.isBorrowProcessing)
-    }
-
-    func testIsBorrowProcessing_IgnoresNotificationsForDifferentBook() {
-        let (vm, _, _) = makeVM()
-        XCTAssertFalse(vm.isBorrowProcessing)
-
-        // Notification for a *different* book ID — must NOT affect this VM.
-        NotificationCenter.default.post(
-            name: .TPPBookProcessingDidChange,
-            object: nil,
-            userInfo: [
-                TPPNotificationKeys.bookProcessingBookIDKey: "some-other-book-id",
-                TPPNotificationKeys.bookProcessingValueKey: true
-            ]
-        )
-
-        // Drain the main run loop so any (incorrect) delivery has a chance to fire.
-        let drain = expectation(description: "drain run loop")
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.15) { drain.fulfill() }
-        wait(for: [drain], timeout: 1.0)
-
-        XCTAssertFalse(
-            vm.isBorrowProcessing,
-            "Processing notifications must be filtered by book identifier — otherwise an unrelated book's borrow puts every detail view into a spinning state."
-        )
-    }
-}
